# release-20.2: sql: implement sequence caching and cached sequence serial normalization