--- a/pkg/sql/catalog/systemschema/system.go
+++ b/pkg/sql/catalog/systemschema/system.go
@@ -628,6 +628,7 @@ var (
 			MinValue:  1,
 			MaxValue:  math.MaxInt64,
 			Start:     1,
+			CacheSize: 1,
 		},
 		Privileges: descpb.NewCustomSuperuserPrivilegeDescriptor(
 			descpb.SystemAllowedPrivileges[keys.DescIDSequenceID], security.NodeUser),