--- a/pkg/sql/logictest/testdata/logic_test/sequences
+++ b/pkg/sql/logictest/testdata/logic_test/sequences
@@ -96,20 +96,11 @@ CREATE SEQUENCE limit_test MAXVALUE 10 START WITH 11
 statement error pgcode 22023 START value \(5\) cannot be less than MINVALUE \(10\)
 CREATE SEQUENCE limit_test MINVALUE 10 START WITH 5
 
-statement error pgcode 22023 CACHE \(-1\) must be greater than zero
-CREATE SEQUENCE cache_test CACHE -1
-
-statement error pgcode 22023 CACHE \(0\) must be greater than zero
-CREATE SEQUENCE cache_test CACHE 0
-
-statement error pgcode 0A000 CACHE values larger than 1 are not supported, found 5
-CREATE SEQUENCE cache_test CACHE 5
-
 statement error pgcode 0A000 CYCLE option is not supported
 CREATE SEQUENCE cycle_test CYCLE
 
 statement ok
-CREATE SEQUENCE ignored_options_test CACHE 1 NO CYCLE
+CREATE SEQUENCE ignored_options_test NO CYCLE
 
 # Verify presence in crdb_internal.create_statements.
 
@@ -120,7 +111,7 @@ query ITTITTTTTTTB colnames
 SELECT * FROM crdb_internal.create_statements WHERE descriptor_name = 'show_create_test'
 ----
 database_id  database_name  schema_name  descriptor_id  descriptor_type  descriptor_name   create_statement                                                                                     state   create_nofks                                                                                         alter_statements  validate_statements  has_partitions
-52           test           public       66             sequence         show_create_test  CREATE SEQUENCE public.show_create_test MINVALUE 1 MAXVALUE 9223372036854775807 INCREMENT 1 START 1  PUBLIC  CREATE SEQUENCE public.show_create_test MINVALUE 1 MAXVALUE 9223372036854775807 INCREMENT 1 START 1  {}                {}                   false
+52           test           public       63             sequence         show_create_test  CREATE SEQUENCE public.show_create_test MINVALUE 1 MAXVALUE 9223372036854775807 INCREMENT 1 START 1  PUBLIC  CREATE SEQUENCE public.show_create_test MINVALUE 1 MAXVALUE 9223372036854775807 INCREMENT 1 START 1  {}                {}                   false
 
 query TT colnames
 SHOW CREATE SEQUENCE show_create_test
@@ -1279,7 +1270,13 @@ ALTER SEQUENCE db2.seq OWNED BY db1.t.a
 statement ok
 CREATE SEQUENCE db2.seq2 OWNED BY db1.t.a
 
+statement error relation "" does not exist
+ALTER SEQUENCE db2.seq2 OWNED BY doesntexist
+
 # Unit test for #60737
+# Test that passing a descriptor other than a sequence
+# returns an appropriate error.
+subtest invalid_ids
 
 statement ok
 CREATE SEQUENCE s1 MINVALUE -4 START WITH -2 CACHE 1 INCREMENT BY -2;
@@ -1329,3 +1326,76 @@ DROP SEQUENCE s1
 statement ok
 DROP TABLE s2
 
+statement error does not exist
+SELECT currval(12345::regclass) # Bogus ID
+
+subtest cached_sequences
+
+statement error pgcode 22023 CACHE \(-1\) must be greater than zero
+CREATE SEQUENCE cache_test CACHE -1
+
+statement error pgcode 22023 CACHE \(0\) must be greater than zero
+CREATE SEQUENCE cache_test CACHE 0
+
+statement ok
+CREATE SEQUENCE cache_test CACHE 10 INCREMENT 1
+
+# Verify cache invalidation with schema changes.
+
+# 10 values (1,2,...,10) are cached, and the underlying sequence is incremented to 10.
+query I
+SELECT nextval('cache_test')
+----
+1
+
+# sanity checks
+query I
+SELECT lastval()
+----
+1
+
+query I
+SELECT currval('cache_test')
+----
+1
+
+query I
+SELECT last_value FROM cache_test
+----
+10
+
+statement ok
+BEGIN
+
+statement ok
+ALTER SEQUENCE cache_test INCREMENT 5
+
+# The cache is invalidated due to the above schema change, and 10 new values (15,20,...,60) are cached.
+query I
+SELECT nextval('cache_test')
+----
+15
+
+# Rollback the schema change to use the old INCREMENT amount.
+statement ok
+ABORT
+
+# The underlying sequence was still incremented despite the txn being aborted.
+query I
+SELECT last_value FROM cache_test
+----
+60
+
+# 10 new values (61,62,...,70) are cached.
+query I
+SELECT nextval('cache_test')
+----
+61
+
+query I
+SELECT last_value FROM cache_test
+----
+70
+
+statement ok
+DROP SEQUENCE cache_test