--- a/pkg/sql/show_create_clauses.go
+++ b/pkg/sql/show_create_clauses.go
@@ -237,6 +237,9 @@ func ShowCreateSequence(
 	if opts.Virtual {
 		f.Printf(" VIRTUAL")
 	}
+	if opts.CacheSize > 1 {
+		f.Printf(" CACHE %d", opts.CacheSize)
+	}
 	return f.CloseAndGetString(), nil
 }