… docs/generated/settings/settings.html
… pkg/sql/catalog/descpb/structured.pb.go
… pkg/sql/sequence_test.go
… pkg/sql/show_test.go