@rafiss 2021-05-05
Richard is out until tomorrow, and I am out tomorrow and Thursday.

I hadn't really looked into the implementation of this before. I don't know of a specific reason not to, apart from it being a new feature so not something that we'd generally backport. Are we supposed to be rigorously following this proposal? https://docs.google.com/document/d/1icfBB-WlzyStvl4t5cTYAV79RQGXvtCHOAvRV7EeC7g/edit

If we are backporting, we'd need this too: https://github.com/cockroachdb/cockroach/pull/63548